Move ROITableWidget into spectrum_widgets/roi_form_widget.py
