+++ mantidimaging/gui/ui/roi_form_widget.ui
@@ -214,7 +214,7 @@
   <customwidget>
    <class>ROITableWidget</class>
    <extends>QTableView</extends>
-   <header>mantidimaging.gui.windows.spectrum_viewer.view</header>
+   <header>mantidimaging.gui.widgets.spectrum_widgets.roi_form_widget</header>
   </customwidget>
   <customwidget>
    <class>ROIPropertiesTableWidget</class>
