+++ mantidimaging/gui/windows/spectrum_viewer/view.py
@@ -3,139 +3,27 @@
 from __future__ import annotations
 
 from pathlib import Path
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 from PyQt5.QtGui import QPixmap
-from PyQt5.QtWidgets import (QCheckBox, QVBoxLayout, QFileDialog, QLabel, QAbstractItemView, QHeaderView, QComboBox,
-                             QSpinBox, QGroupBox, QActionGroup, QAction)
-from PyQt5.QtCore import QModelIndex, pyqtSignal
+from PyQt5.QtWidgets import (QCheckBox, QVBoxLayout, QFileDialog, QLabel, QComboBox, QSpinBox, QGroupBox, QActionGroup,
+                             QAction)
+from PyQt5.QtCore import QModelIndex
 
 from mantidimaging.core.utility import finder
 from mantidimaging.gui.mvp_base import BaseMainWindowView
 from mantidimaging.gui.widgets.dataset_selector import DatasetSelectorWidgetView
 from .model import ROI_RITS, allowed_modes
 from .presenter import SpectrumViewerWindowPresenter, ExportMode
-from mantidimaging.gui.widgets import RemovableRowTableView
 from .spectrum_widget import SpectrumWidget
-from mantidimaging.gui.windows.spectrum_viewer.roi_table_model import TableModel
 from mantidimaging.gui.widgets.spectrum_widgets.tof_properties import ExperimentSetupFormWidget
 from mantidimaging.gui.widgets.spectrum_widgets.roi_selection_widget import ROISelectionWidget
 import numpy as np
 
 if TYPE_CHECKING:
     from mantidimaging.gui.windows.main import MainWindowView  # noqa:F401  # pragma: no cover
-    from mantidimaging.gui.widgets.spectrum_widgets.roi_form_widget import ROIFormWidget
+    from mantidimaging.gui.widgets.spectrum_widgets.roi_form_widget import ROIFormWidget, ROITableWidget
     from uuid import UUID
-
-
-class ROITableWidget(RemovableRowTableView):
-    """
-    A class to represent the ROI table widget in the spectrum viewer window.
-    """
-    ElementType = str | tuple[int, int, int] | bool
-    RowType = list[ElementType]
-    selected_row: int
-
-    selection_changed = pyqtSignal()
-    name_changed = pyqtSignal(str, str)
-    visibility_changed = pyqtSignal()
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-
-        self.selected_row = 0
-
-        # Point table
-        self.setSelectionBehavior(QAbstractItemView.SelectRows)
-        self.setSelectionMode(QAbstractItemView.SingleSelection)
-        self.setAlternatingRowColors(True)
-
-        # Initialise model
-        self.roi_table_model = TableModel()
-        self.setModel(self.roi_table_model)
-
-        # Configure up the table view
-        self.setVisible(True)
-        header = self.horizontalHeader()
-        header.setSectionResizeMode(0, QHeaderView.Stretch)
-        header.setSectionResizeMode(1, QHeaderView.ResizeToContents)
-        header.setSectionResizeMode(2, QHeaderView.ResizeToContents)
-
-        self.selectionModel().currentRowChanged.connect(self.on_row_change)
-        self.roi_table_model.name_changed.connect(self.name_changed.emit)
-        self.roi_table_model.visibility_changed.connect(self.visibility_changed.emit)
-
-    def on_row_change(self, item: QModelIndex, _: Any) -> None:
-        self.selected_row = item.row()
-        self.selection_changed.emit()
-
-    def get_roi_names(self) -> list[str]:
-        return self.roi_table_model.roi_names()
-
-    def get_roi_name_by_row(self, row: int) -> str:
-        """
-        Retrieve the name an ROI by its row index.
-        """
-        return self.roi_table_model.get_element(row, 0)
-
-    @property
-    def current_roi_name(self) -> str:
-        return self.get_roi_name_by_row(self.selected_row)
-
-    def get_roi_visibility_by_row(self, row: int) -> bool:
-        """
-        Retrieve the visibility status of an ROI by its row index.
-        """
-        return self.roi_table_model.get_element(row, 2)
-
-    def row_count(self) -> int:
-        """
-        Returns the number of rows in the ROI table model.
-        """
-        return self.roi_table_model.rowCount()
-
-    def find_row_for_roi(self, roi_name: str) -> int | None:
-        """
-        Returns row index for ROI name, or None if not found.
-        """
-        for row in range(self.roi_table_model.rowCount()):
-            if self.roi_table_model.index(row, 0).data() == roi_name:
-                return row
-        return None
-
-    def update_roi_color(self, roi_name: str, new_color: tuple[int, int, int, int]) -> None:
-        """
-        Finds ROI by name in table and updates it's colour (R, G, B) format.
-        """
-        row = self.find_row_for_roi(roi_name)
-        if row is not None:
-            self.roi_table_model.update_color(row, new_color)
-
-    def add_row(self, name: str, colour: tuple[int, int, int, int], roi_names: list[str]) -> None:
-        """
-        Add a new row to the ROI table
-        """
-        self.roi_table_model.appendNewRow(name, colour, True)
-        self.selected_row = self.roi_table_model.rowCount() - 1
-        self.selectRow(self.selected_row)
-
-    def remove_row(self, row: int) -> None:
-        """
-        Remove a row from the ROI table
-        """
-        self.roi_table_model.remove_row(row)
-        self.selectRow(0)
-
-    def clear_table(self) -> None:
-        """
-        Clears the ROI table in the spectrum viewer.
-        """
-        self.roi_table_model.clear_table()
-
-    def select_roi(self, roi_name: str) -> None:
-        selected_row = self.find_row_for_roi(roi_name)
-        assert selected_row is not None
-        self.selectRow(selected_row)
 
 
 class SpectrumViewerWindowView(BaseMainWindowView):
